requirements.txt
Django==1.9.6
djangorestframework==3.3.3
coverage==4.0.3
twine==1.6.5
pylint==1.5.5
pep8==1.7.0
